Increased mission start time and fixed Tembelan weather issue
To fully load in the prison consistently I had to set the sleep value to 12.  It can load faster, but it depends on the prison.

Tembelan island also had an issue where the map would load as incredible foggy a significant portion of the time regardless of the weather parameters.  Removing the "startFogBase=250;" from the mission.sqm fixes that issue.