New CUP RU vs US Army-Taki GUE and CUP SLA vs HIL-RACS versions
With the latest CUP update adding a new faction and a bunch of changes to the US Army/USMC and the previous one adding a ton of vehicles to the Taki GUE faction I decided to make some new versions to take advantage of the new gear.

The CUP RU vs US Army-Taki GUE has the players using the new Ratnik pink desert uniforms mostly because they're so unique.  The main opposing faction is the US Army and the secondary is the Takistan Guerrillas.  It has a lot of weapon variety and it's interesting to see such contrast between the opposing factions, especially with wartorn on. I've added this version to desert maps.

The CUP SLA vs HIL-RACS has the players as the Sahrani Liberation Army with the default uniforms against the new Horizon Islands Legion with their default uniforms and the Royal Army Corps of Sahrani with their jungle uniforms.  I've added this version to jungle/tropical maps.  The HIL bring a lot of new assets including the FAMAS and a new APC and Tank so people who play escape for the variety of enemies (like me) will be pleased.