Ditch compile_local.py and minor improvement to run locally
I just started with working with the source of this mission and tried to
build it locally. However, the builddir and packed dir where not
created. One of the improvements here is now, that these directories are
created automatically if they are not present.

Also: When looking at the compile.py and compile_local.py, they loocked
pretty much identical. I merged the different things and made the
compile.py script being able to run locally and therefore
compile_local.py could be removed (which I boldly did in this PR as
well).