Weather System Overhaul
The weather system in Escape for a while has had a lot of issues, as detailed in my bug report (https://github.com/NeoArmageddon/co10_Escape/issues/20), but to put it short the main issues are the startup parameters basically don't work at all, and the second issue is that dynamic weather doesn't really work properly in Arma, with clounds frequently rendering improperly creating a strange look for players.

I've decided to replace the current system with a static, in engine system with loads and loads of parameters for players to tweak to their hearts content.  Want to have a bright sunny day with just a touch of fog?  Check.  Want to do a dark and stormy night?  Check.  Want to put the wind value to 100m/s and watch the clouds move like they're running away from something?  Check.  I was able to keep a small amount of in mission weather changes since by default in the engine rain and wind adjust dynamically during a mission based on initial conditions (mostly overcast it seems) and the player can choose to use keep them static or let the engine decide.

I chose not to add settings for waves and lightning since the manual overide setting needs to be toggled in the editor to work which would require an adjustment to every mission.sqm and those settings are already accounted for in the engine anyways.  For high waves simply set high wind and for high lightning put high rain (not as sure on lightning since it's effect is less dramatic).

To prevent any conflicts I've stripped out all the old weather scripts and all the calls to them I could find.  If there's any I missed let me know and I'll edit them out too.

For an unrelated change I also added a small change to the description for the search helicopters to better explain how it works.  Figured I'd put it in here since I'm already adjusting the params.hpp file.